refactor: extract usePlyrPlayer composable from VideoDialog (3a)
Player lifecycle, load-id race guard, position restore, transcript
control injection, and the fullscreen Escape capture move to
composables/usePlyrPlayer.ts behind a small API (loadPlayer,
captureResume/resetResume, seekTo, stop).
