+++ app/components/VideoDialog.vue
@@ -133,36 +133,30 @@
 </template>
 
 <script setup lang="ts">
-import type { Track } from 'plyr';
-import type { MediaFile, Video } from '~/types';
+import type { Video } from '~/types';
 import { useDisplay } from 'vuetify';
 
 const store = useAppStore();
 const route = useRoute();
 const router = useRouter();
 const { xs, smAndDown } = useDisplay();
-const { isCastConnected, isMediaLoaded, currentTime: castTime, seekTo } = useCast();
+const { isCastConnected, isMediaLoaded, currentTime: castTime, seekTo: castSeekTo } = useCast();
 
 const playerEl = ref<HTMLVideoElement | null>(null);
-let player: Plyr | undefined = undefined;
-let playerLoadId = 0;
-// Playback position to restore after a language switch rebuilds the player
-let resumeTime = 0;
 
 const loading = ref(true);
 const videoMedia = ref<Video | null>(null);
 const subtitleMedia = ref<Video | null>(null);
-const localTime = ref(0);
 
 const isCasting = computed(() => isCastConnected.value && isMediaLoaded.value);
 const transcriptTime = computed(() => (isCasting.value ? castTime.value : localTime.value));
 
 function onSeekTranscript(seconds: number) {
   if (isCasting.value) {
-    seekTo(seconds);
+    castSeekTo(seconds);
   }
-  else if (player) {
-    player.currentTime = seconds;
+  else {
+    playerSeekTo(seconds);
   }
 }
 
@@ -223,6 +217,15 @@ const availableLanguages = computed(() => {
   );
 });
 
+const {
+  localTime,
+  loadPlayer,
+  captureResume,
+  resetResume,
+  seekTo: playerSeekTo,
+  stop: stopPlayer,
+} = usePlyrPlayer(playerEl, { videoMedia, captionUrl, subtitleUrl, poster: videoPoster });
+
 async function loadMediaItems() {
   if (!store.selectedVideo) {
     return;
@@ -254,140 +257,6 @@ async function loadMediaItems() {
   loading.value = false;
 }
 
-let transcriptBtn: HTMLButtonElement | null = null;
-
-// mdi script-text-outline — same icon as ButtonTranscript
-const transcriptIconPath = 'M15,20A1,1 0 0,0 16,19V4H8A1,1 0 0,0 7,5V16H5V5A3,3 0 0,1 8,2H19A3,3 0 0,1 22,5V6H20V5A1,1 0 0,0 19,4A1,1 0 0,0 18,5V9L18,19A3,3 0 0,1 15,22H5A3,3 0 0,1 2,19V18H13A2,2 0 0,0 15,20M9,6H14V8H9V6M9,10H14V12H9V10M9,14H14V16H9V14Z';
-
-function syncTranscriptButton() {
-  transcriptBtn?.setAttribute('aria-pressed', String(store.transcriptDialog));
-}
-
-// Plyr has no API for extra controls; insert a button into its control bar.
-// Runs on every 'ready' because assigning player.source rebuilds the controls DOM.
-function injectTranscriptButton() {
-  if (smAndDown.value || !player) {
-    return;
-  }
-  const controls = player.elements.controls;
-  if (!controls || controls.querySelector('.plyr__control--transcript')) {
-    return;
-  }
-  const btn = document.createElement('button');
-  btn.type = 'button';
-  btn.className = 'plyr__controls__item plyr__control plyr__control--transcript';
-  btn.innerHTML
-    = `<svg aria-hidden="true" focusable="false" viewBox="0 0 24 24">`
-      + `<path fill="currentColor" d="${transcriptIconPath}"/></svg>`
-      + '<span class="plyr__sr-only">Transcript</span>';
-  btn.addEventListener('click', () => store.setTranscriptDialog(!store.transcriptDialog));
-  const fullscreenBtn = player.elements.buttons.fullscreen;
-  if (fullscreenBtn && fullscreenBtn.parentElement === controls) {
-    fullscreenBtn.before(btn);
-  }
-  else {
-    controls.append(btn);
-  }
-  transcriptBtn = btn;
-  syncTranscriptButton();
-}
-
-watch(() => store.transcriptDialog, syncTranscriptButton);
-
-// Escape while in fullscreen should only exit fullscreen (browser default),
-// not also bubble into Vuetify's overlay and close the whole dialog
-function onKeydownCapture(e: KeyboardEvent) {
-  if (e.key === 'Escape' && document.fullscreenElement) {
-    e.stopImmediatePropagation();
-  }
-}
-onMounted(() => document.addEventListener('keydown', onKeydownCapture, true));
-onBeforeUnmount(() => document.removeEventListener('keydown', onKeydownCapture, true));
-
-async function loadPlayer() {
-  if (!playerEl.value || !videoMedia.value) {
-    return;
-  }
-
-  // The dialog-open and loading watchers can both request an init in the same
-  // tick; without this guard two calls interleave across the dynamic import
-  // and the second destroys the first mid-setup, leaving a dead player
-  const id = ++playerLoadId;
-  const { default: Plyr } = await import('plyr');
-  if (id !== playerLoadId || !playerEl.value || !videoMedia.value) {
-    return;
-  }
-  if (player) {
-    player.destroy();
-  }
-
-  const tracks: Track[] = [];
-  if (captionUrl.value) {
-    tracks.push({
-      kind: 'captions',
-      label: languageLabel(store.getVideoLanguage!),
-      srcLang: store.getVideoLanguage!.locale,
-      src: captionUrl.value,
-    });
-  }
-  if (subtitleUrl.value) {
-    tracks.push({
-      kind: 'subtitles',
-      label: languageLabel(store.getSubtitleLanguage!),
-      srcLang: store.getSubtitleLanguage!.locale,
-      src: subtitleUrl.value,
-    });
-  }
-
-  player = new Plyr(playerEl.value, {
-    quality: { default: 1080, options: [1080, 720, 480, 360, 240] },
-    captions: { active: true, language: store.getSubtitleLanguage!.locale, update: true },
-    // Few enough options that the settings menu can't soft-lock
-    speed: { selected: 1, options: [0.75, 1, 1.25, 1.5] },
-    // Fullscreen the whole row so the transcript panel stays visible (desktop)
-    ...(smAndDown.value ? {} : { fullscreen: { container: '.player-row' } }),
-  });
-  player.on('ready', injectTranscriptButton);
-
-  localTime.value = 0;
-  player.on('timeupdate', () => {
-    localTime.value = player?.currentTime ?? 0;
-  });
-
-  player.source = {
-    type: 'video',
-    poster: videoPoster.value,
-    title: store.selectedVideo?.title,
-    sources:
-      videoMedia.value.files.map((f: MediaFile) => ({
-        src: f.progressiveDownloadURL,
-        type: f.mimetype,
-        size: Number.parseInt(f.label.slice(0, -1), 10),
-      })) ?? [],
-    tracks,
-  };
-
-  const resume = resumeTime;
-  resumeTime = 0;
-  if (resume > 0) {
-    const instance = player;
-    // Plyr's source setter swaps in a fresh media element asynchronously, so
-    // seek only once it re-emits 'ready'. Set the element's currentTime
-    // directly (Plyr's own setter bails while duration is still unknown) —
-    // before metadata this records the default playback start position —
-    // and re-assert on loadedmetadata once the real duration is in.
-    instance.once('ready', () => {
-      const media = instance.elements.container?.querySelector('video');
-      if (media) {
-        media.currentTime = resume;
-      }
-    });
-    instance.once('loadedmetadata', () => {
-      instance.currentTime = resume;
-    });
-  }
-}
-
 // Re-init player once media is loaded
 watch(loading, async isLoading => {
   if (!isLoading) {
@@ -401,7 +270,7 @@ watch(
   () => store.videoDialog,
   open => {
     if (!open) {
-      player?.stop();
+      stopPlayer();
       store.setTranscriptDialog(false);
       const lang = route.params.language as string;
       if (route.params.videoId) {
@@ -429,7 +298,7 @@ watch(
       return;
     }
     store.setTranscriptDialog(false);
-    resumeTime = 0;
+    resetResume();
     videoMedia.value = null;
     subtitleMedia.value = null;
     // Pre-fill from selectedVideo if language matches
@@ -447,7 +316,7 @@ watch(
 watch(
   () => store.videoLanguage,
   () => {
-    resumeTime = player?.currentTime ?? localTime.value;
+    captureResume();
     videoMedia.value = null;
     loadMediaItems();
   },
@@ -457,7 +326,7 @@ watch(
 watch(
   () => store.subtitleLanguage,
   () => {
-    resumeTime = player?.currentTime ?? localTime.value;
+    captureResume();
     subtitleMedia.value = null;
     loadMediaItems();
   },
