+++ app/composables/usePlyrPlayer.ts
@@ -1,0 +1,183 @@
+import type Plyr from 'plyr';
+import type { Track } from 'plyr';
+import type { Ref } from 'vue';
+import type { Video } from '~/types';
+import { useDisplay } from 'vuetify';
+
+// mdi script-text-outline — same icon as the transcript button
+const TRANSCRIPT_ICON_PATH = 'M15,20A1,1 0 0,0 16,19V4H8A1,1 0 0,0 7,5V16H5V5A3,3 0 0,1 8,2H19A3,3 0 0,1 22,5V6H20V5A1,1 0 0,0 19,4A1,1 0 0,0 18,5V9L18,19A3,3 0 0,1 15,22H5A3,3 0 0,1 2,19V18H13A2,2 0 0,0 15,20M9,6H14V8H9V6M9,10H14V12H9V10M9,14H14V16H9V14Z';
+
+/**
+ * Owns the Plyr lifecycle for the video dialog: (re-)initialization behind a
+ * load-id race guard, playback-position restore across rebuilds, the injected
+ * transcript control, and the fullscreen Escape capture.
+ */
+export function usePlyrPlayer(
+  playerEl: Ref<HTMLVideoElement | null>,
+  source: {
+    videoMedia: Ref<Video | null>;
+    captionUrl: Ref<string | null>;
+    subtitleUrl: Ref<string | null>;
+    poster: Ref<string | undefined>;
+  },
+) {
+  const store = useAppStore();
+  const { smAndDown } = useDisplay();
+
+  let player: Plyr | undefined = undefined;
+  let playerLoadId = 0;
+  // Playback position to restore after a language switch rebuilds the player
+  let resumeTime = 0;
+  const localTime = ref(0);
+
+  let transcriptBtn: HTMLButtonElement | null = null;
+
+  function syncTranscriptButton() {
+    transcriptBtn?.setAttribute('aria-pressed', String(store.transcriptDialog));
+  }
+
+  // Plyr has no API for extra controls; insert a button into its control bar.
+  // Runs on every 'ready' because assigning player.source rebuilds the controls DOM.
+  function injectTranscriptButton() {
+    if (smAndDown.value || !player) {
+      return;
+    }
+    const controls = player.elements.controls;
+    if (!controls || controls.querySelector('.plyr__control--transcript')) {
+      return;
+    }
+    const btn = document.createElement('button');
+    btn.type = 'button';
+    btn.className = 'plyr__controls__item plyr__control plyr__control--transcript';
+    btn.innerHTML
+      = `<svg aria-hidden="true" focusable="false" viewBox="0 0 24 24">`
+        + `<path fill="currentColor" d="${TRANSCRIPT_ICON_PATH}"/></svg>`
+        + '<span class="plyr__sr-only">Transcript</span>';
+    btn.addEventListener('click', () => store.setTranscriptDialog(!store.transcriptDialog));
+    const fullscreenBtn = player.elements.buttons.fullscreen;
+    if (fullscreenBtn && fullscreenBtn.parentElement === controls) {
+      fullscreenBtn.before(btn);
+    }
+    else {
+      controls.append(btn);
+    }
+    transcriptBtn = btn;
+    syncTranscriptButton();
+  }
+
+  watch(() => store.transcriptDialog, syncTranscriptButton);
+
+  // Escape while in fullscreen should only exit fullscreen (browser default),
+  // not also bubble into Vuetify's overlay and close the whole dialog
+  function onKeydownCapture(e: KeyboardEvent) {
+    if (e.key === 'Escape' && document.fullscreenElement) {
+      e.stopImmediatePropagation();
+    }
+  }
+  onMounted(() => document.addEventListener('keydown', onKeydownCapture, true));
+  onBeforeUnmount(() => document.removeEventListener('keydown', onKeydownCapture, true));
+
+  async function loadPlayer() {
+    if (!playerEl.value || !source.videoMedia.value) {
+      return;
+    }
+
+    // The dialog-open and loading watchers can both request an init in the same
+    // tick; without this guard two calls interleave across the dynamic import
+    // and the second destroys the first mid-setup, leaving a dead player
+    const id = ++playerLoadId;
+    const { default: Plyr } = await import('plyr');
+    if (id !== playerLoadId || !playerEl.value || !source.videoMedia.value) {
+      return;
+    }
+    if (player) {
+      player.destroy();
+    }
+
+    const tracks: Track[] = [];
+    if (source.captionUrl.value) {
+      tracks.push({
+        kind: 'captions',
+        label: languageLabel(store.getVideoLanguage!),
+        srcLang: store.getVideoLanguage!.locale,
+        src: source.captionUrl.value,
+      });
+    }
+    if (source.subtitleUrl.value) {
+      tracks.push({
+        kind: 'subtitles',
+        label: languageLabel(store.getSubtitleLanguage!),
+        srcLang: store.getSubtitleLanguage!.locale,
+        src: source.subtitleUrl.value,
+      });
+    }
+
+    player = new Plyr(playerEl.value, {
+      quality: { default: 1080, options: [1080, 720, 480, 360, 240] },
+      captions: { active: true, language: store.getSubtitleLanguage!.locale, update: true },
+      // Few enough options that the settings menu can't soft-lock
+      speed: { selected: 1, options: [0.75, 1, 1.25, 1.5] },
+      // Fullscreen the whole row so the transcript panel stays visible (desktop)
+      ...(smAndDown.value ? {} : { fullscreen: { container: '.player-row' } }),
+    });
+    player.on('ready', injectTranscriptButton);
+
+    localTime.value = 0;
+    player.on('timeupdate', () => {
+      localTime.value = player?.currentTime ?? 0;
+    });
+
+    player.source = {
+      type: 'video',
+      poster: source.poster.value,
+      title: store.selectedVideo?.title,
+      sources: source.videoMedia.value.files.map(f => ({
+        src: f.progressiveDownloadURL,
+        type: f.mimetype,
+        size: Number.parseInt(f.label.slice(0, -1), 10),
+      })),
+      tracks,
+    };
+
+    const resume = resumeTime;
+    resumeTime = 0;
+    if (resume > 0) {
+      const instance = player;
+      // Plyr's source setter swaps in a fresh media element asynchronously, so
+      // seek only once it re-emits 'ready'. Set the element's currentTime
+      // directly (Plyr's own setter bails while duration is still unknown) —
+      // before metadata this records the default playback start position —
+      // and re-assert on loadedmetadata once the real duration is in.
+      instance.once('ready', () => {
+        const media = instance.elements.container?.querySelector('video');
+        if (media) {
+          media.currentTime = resume;
+        }
+      });
+      instance.once('loadedmetadata', () => {
+        instance.currentTime = resume;
+      });
+    }
+  }
+
+  // Remember the current position so the next loadPlayer restores it
+  function captureResume() {
+    resumeTime = player?.currentTime ?? localTime.value;
+  }
+
+  function resetResume() {
+    resumeTime = 0;
+  }
+
+  function seekTo(seconds: number) {
+    if (player) {
+      player.currentTime = seconds;
+    }
+  }
+
+  function stop() {
+    player?.stop();
+  }
+
+  return { localTime, loadPlayer, captureResume, resetResume, seekTo, stop };
+}
